+++ eva/src/js/classes/ConstructorSchemes/elementTemplates.js
@@ -460,7 +460,7 @@ const templates = {
               writingMode: tag.isVertical ? 'vertical-rl' : 'horizontal-tb',
               textOrientation: 'mixed', 
               transform: tag.isVertical ? 'rotate(180deg)' : 'rotate(0deg)',
-              color: tag.textColor,
+              color: tag.textColor.rgbaString,
               fontSize: tag.fontSize ? tag.fontSize + 'px' : '12px',
               fontFamily: tag.fontFamily,
              }"
@@ -485,9 +485,33 @@ const templates = {
       borderType: 'solid',
       borderSize: 1,
       borderDashed: true,
-      borderColor: '#000000',
-      bgColor: '#FFFFFF',
-      textColor: '#000000',
+      borderColor: {
+        rgbaObject: {
+          r: 0,
+          g: 0,
+          b: 0,
+          a: 1,
+        },
+        rgbaString: 'rgba(0,0,0,1)',
+      },
+      bgColor: {
+        rgbaObject: {
+          r: 255,
+          g: 255,
+          b: 255,
+          a: 1,
+        },
+        rgbaString: 'rgba(255,255,255,1)',
+      },
+      textColor: {
+        rgbaObject: {
+          r: 0,
+          g: 0,
+          b: 0,
+          a: 1,
+        },
+        rgbaString: 'rgba(0,0,0,1)',
+      },
       fontSize: 12,
     },
   },