+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Constructor-schemes: the ability to remove a scheme from the list of saved schemes
 - Constructor-schemes: token-action click:image
-- Constructor-schemes: adds default shape - triangle(top)\triangle(bottom)
+- Constructor-schemes: import\export schemes
 
 ### Changed